pyproject.toml
[project]
name = "a-quant"
version = "0.1.0"
description = "A股量化交易项目"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tushare>=1.2.0",
    "akshare>=1.12.0",
    "baostock>=0.8.8",
    "backtrader>=1.9.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/quant"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]

